Migrate to python3.6
